Account for inflight when doing optimistic rollback

+++ money-transfer/src/money_transfer/optimistic_locking_and_rollback.py
@@ -14,33 +14,74 @@ if TYPE_CHECKING:
 
 def current_balance(ctx: Context, account_id: int) -> tuple[int, int]:
     conn: Connection = ctx.deps.get("conn")
-    balance, version = conn.execute(
-        "SELECT balance, version FROM accounts WHERE account_id = ?",
-        (account_id,),
-    ).fetchone()
-    conn.commit()
+    with conn:
+        balance_with_version: tuple[int, int] = conn.execute(
+            """
+                SELECT balance - IFNULL(
+                    (
+                    SELECT SUM(amount)
+                    FROM transfers
+                    WHERE account_id = ?
+                    AND amount > 0
+                    ), 0)
+                , version
+                FROM accounts
+                WHERE account_id = ?
+                """,
+            (
+                account_id,
+                account_id,
+            ),
+        ).fetchone()
+        balance, version = balance_with_version
     return balance, version
 
 
 def rollback(ctx: Context, transfer_id: int) -> None:
     conn: Connection = ctx.deps.get("conn")
+    with conn:
+        transfers: list[tuple[int, int]] = conn.execute(
+            "SELECT amount, account_id FROM transfers WHERE transfer_id = ?",
+            (transfer_id,),
+        ).fetchall()
 
-    transfer: list[tuple[int, int]] = conn.execute(
-        "SELECT account_id, amount FROM transfers WHERE transfer_id = ?", (transfer_id,)
-    ).fetchall()
-    for account_id, amount in transfer:
+        if len(transfers) > 0:
+            cur = conn.executemany(
+                """
+                    UPDATE accounts
+                    SET
+                        balance = balance - ?,
+                        version = version + 1
+                    WHERE account_id = ?
+                    """,
+                transfers,
+            )
+            ctx.assert_statement(cur.rowcount > 0, "More that one row must be affected")
+            conn.execute("DELETE FROM transfers WHERE transfer_id = ?", (transfer_id,))
+
+
+def update_balance_ensuring_version(
+    ctx: Context, transfer_id: int, account_id: int, amount: int, version: int
+) -> None:
+    conn: Connection = ctx.deps.get("conn")
+    with conn:
         conn.execute(
-            """
-            UPDATE accounts
-            SET
-                balance = balance - ?,
-                version = version + 1
-            WHERE account_id = ?
-            """,
-            (amount, account_id),
+            "INSERT INTO transfers (transfer_id, account_id, amount) VALUES (?, ?, ?)",
+            (transfer_id, account_id, amount),
         )
-    conn.execute("DELETE FROM transfers WHERE transfer_id = ?", (transfer_id,))
-    conn.commit()
+        cur = conn.execute(
+            """
+                UPDATE accounts
+                SET
+                    balance = balance + ?,
+                    version = version + 1
+                WHERE account_id = ? and version = ?
+                """,
+            (amount, account_id, version),
+        )
+        if cur.rowcount == 0:
+            raise errors.VersionConflictError
+    ctx.assert_statement(cur.rowcount == 1, msg="More that one row was affected")
 
 
 def update_balance(
@@ -48,42 +89,24 @@ def update_balance(
     transfer_id: int,
     account_id: int,
     amount: int,
-    with_version: int | None = None,
 ) -> None:
     conn: Connection = ctx.deps.get("conn")
+    with conn:
+        conn.execute(
+            "INSERT INTO transfers (transfer_id, account_id, amount) VALUES (?, ?, ?)",
+            (transfer_id, account_id, amount),
+        )
 
-    conn.execute(
-        "INSERT INTO transfers (transfer_id, account_id, amount) VALUES (?, ?, ?)",
-        (transfer_id, account_id, amount),
-    )
-    if with_version is None:
         cur = conn.execute(
             """
-            UPDATE accounts
-            SET
-                balance = balance + ?,
-                version = version + 1
-            WHERE account_id = ?
-            """,
+                UPDATE accounts
+                SET
+                    balance = balance + ?,
+                    version = version + 1
+                WHERE account_id = ?
+                """,
             (amount, account_id),
         )
-
-    else:
-        cur = conn.execute(
-            """
-            UPDATE accounts
-            SET
-                balance = balance + ?,
-                version = version + 1
-            WHERE account_id = ? and version = ?
-            """,
-            (amount, account_id, with_version),
-        )
-        if cur.rowcount == 0:
-            conn.rollback()
-            raise errors.VersionConflictError
-
-    conn.commit()
 
     ctx.assert_statement(cur.rowcount == 1, msg="More that one row was affected")
 
@@ -95,6 +118,9 @@ def transaction(
     target: int,
     amount: int,
 ) -> Generator[Yieldable, Any, tuple[int, int, int]]:
+    if source == target:
+        raise errors.SameAccountTransferError
+
     yield ctx.call(rollback, transfer_id=transfer_id)
 
     source_balance_and_version: tuple[int, int] = yield ctx.call(
@@ -107,17 +133,17 @@ def transaction(
         raise errors.NotEnoughFundsError(account_id=source)
 
     yield ctx.call(
-        update_balance,
+        update_balance_ensuring_version,
         transfer_id=transfer_id,
         account_id=source,
         amount=amount * -1,
-        with_version=version,
+        version=version,
     )
-
     yield ctx.call(
         update_balance,
         transfer_id=transfer_id,
         account_id=target,
-        amount=amount * 1,
+        amount=amount,
     )
+
     return source, target, amount
